[ty] Move some dynamic class code out of builder.rs
