+++ crates/ty_python_semantic/src/types/infer/builder/new_class.rs
@@ -1,4 +1,3 @@
-use super::{ArgumentsIter, DynamicClassKind, TypeInferenceBuilder};
 use crate::semantic_index::definition::Definition;
 use crate::types::class::{
     ClassLiteral, DynamicClassAnchor, DynamicClassLiteral, DynamicMetaclassConflict,
@@ -7,6 +6,10 @@ use crate::types::class::{
 use crate::types::diagnostic::{
     INVALID_ARGUMENT_TYPE, NO_MATCHING_OVERLOAD, report_conflicting_metaclass_from_bases,
     report_instance_layout_conflict,
+};
+use crate::types::infer::builder::{
+    ArgumentsIter, TypeInferenceBuilder,
+    dynamic_class::{DynamicClassKind, report_dynamic_mro_errors},
 };
 use crate::types::{KnownClass, SubclassOfType, Type, TypeContext, definition_expression_type};
 use ruff_python_ast::{self as ast, HasNodeIndex, NodeIndex};
@@ -156,8 +159,7 @@ impl<'db> TypeInferenceBuilder<'db, '_> {
                 DynamicClassKind::NewClass,
             );
 
-            if super::report_dynamic_mro_errors(&self.context, dynamic_class, call_expr, bases_arg)
-            {
+            if report_dynamic_mro_errors(&self.context, dynamic_class, call_expr, bases_arg) {
                 // MRO succeeded, check for instance-layout-conflict.
                 disjoint_bases.remove_redundant_entries(db);
                 if disjoint_bases.len() > 1 {
